Rearranaged the beamline eventcut files.
After checking for differences between the three previous run ranges,
9000-12162, 13670-15401, and 15402-, I found that the only meaningful
differences in the cuts were in the charge cuts.
I have removed all three of these run ranged files.

There are now two beamline eventcut files:  the unranged file,
which is appropriate for Run2, and run range 1-12170 for Run1.
The only difference between these two files are the presence
of BCM7 and BCM8 in the unranged file.

What are the cuts:
  *  The beam current cut is done on qwk_charge; these files
     have a minimum qwk_charge level of 100 uA.
     The stability cut is on qwk_charge.
  *  All BCMs contribute to the global cut, with a very loose
     current cut range and no stability cut.
  *  All BPM wires have local cuts.
  *  The BPMs used in regression (3C12, 3H04, 3H07A/B/C, 3H09,
     3H09B) have global cuts on the positions with no stability
     cut.
  *  The target BPM does have a stability cut.
  *  The qwk_energy variable contributes to the global cut with
     a loose cut range and no stability cut.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4695 b9700a98-9937-0410-b777-f318d6eb5bff