NOTE:  this is on branch release_3.02.
We have modified the working path for the linRegBlue scripts
and configuration directories, to recover the run-ranged
configurations in which certain data elements are removed
from the regression.

The QwPass5 results to date have used a single set of LRB
configuration files, equivalent to the "run1pass5" directory.
This is our standard set of LRB configurations, with no
channels disabled.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.02@4739 b9700a98-9937-0410-b777-f318d6eb5bff