This script creates the graphs of interst when it comes to light weighting. These are - light yield vs. position on bar (in cm) (both histogram, profile and table) - Q2 vs position on bar (both histogram, porfile,  and table) - Scattering angle  vs position on bar (both histogram, profile, and table) - Q2 w/ light weighting w/o pedestal - Q2 w/o light weighitng - Q2 w/ light wieghitn and pedestal
It also creates a text file that outputs the number of electrons that goes through a part of the bar, 
the Q2 value and error for a part of the bar, 
the light weighting and error for a part of the bar, and
the scattering angle and error for a part of the bar.

NOTE: All octants are rotated so that they are in octant 3.  This means that the 
position on the bar is of X in this octant correspoding to the long side of all the bars.
 


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.01@4755 b9700a98-9937-0410-b777-f318d6eb5bff