After blinding the helicity-correlated differences and just before computing the helicity-correlated asymmetries, update the global error codes for the "difference" and "yield" objects.
This ensure that blinding errors are treated as global for all
yields, asymmetries (HC and alternate), and differences when they
are averaged by the engine.
This also ensures that the Hel_Tree "ErrorFlag" is non-zero when
there are blinding errors in patterns which are otherwise good.

Note that which blinding fails, all blinded asymmetries are set
to zero.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4771 b9700a98-9937-0410-b777-f318d6eb5bff