QwSimTracking, the tracking analyzer that takes QwGeant4 files as input, is now working again (somewhat).
When you run QwSimTracking over a QwSim_${run}.root file using qwsimtracking -r ${run} you get a ROOT file
that has the usual event_tree, but in addition to the branch "events" there is now a branch "originals".
The branch "events" is run through tracking after smearing of the hits by their resolution.  The branch
"originals" is the exact info from the simulation.

For example, event_tree->Draw("events.fQwPartialTracks.fSlopeX - originals.fQwPartialTracks.fSlopeX","...")
will generate a histogram of the difference between actual and reconstructed horizontal slope of the partial
tracks, and therefore an estimate of the uncertainty introduced.  Because some events don't get reconstructed
you need to supply the appropriate ... cuts, or make this part of a ROOT macro.

Todo:
- Pass the cross section weighting factor for comparisons with actual data.
- Pass kinematic variables in originals.
- Fix region 2 in QwSimTracking (currently not filled, for some reason)
  \- Reconstruct momentum in QwSimTracking
- ...


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4795 b9700a98-9937-0410-b777-f318d6eb5bff