Fixed the QwSimTracking functionality.  Output files from QwGeant4 can be run through the tracking analyzer again.
This fixed involved among other things:
- figuring out that after an undocumented commit hits with hit number > 0 are thrown away
  (we need to determine if this is OK for data)
- reconstruct full event in QwTreeEventBuffer, just like for real data
- changed pointers to shared_ptr for reduced memory leakage (I'm a lazy^H^H^H^Hefficient programmer)
- commented some code, removed some commented code

The exact simulated data is in 'originals', the reconstructed events in 'events'.  Originals from QwGeant4 are
smeared by the detector resolution (from geometry file).  QwGeant4 geometry of HDC and VDC is based of the tracking
analyzer geometry file, which is then used to reconstruct events, so geometry should be self-consistent (see
QwDumpGeant4Geometry for details).

Todo:
- Octant support: QwGeant4 has HDC and VDC fixed in octant 3, QwSimTracking rotates that all in octant 5, this means
  that the small geometry difference between 3 and 5 are present between events and originals. Octants are also not
  written to the QwGeant4 output file, which will be necessary just as for Cerenkov events.
- Add tree lines to the originals branch, only partial tracks are there now.
- Add kinematics to the originals branch.

Now the fun with this begins :-\  For example, X and Y are mismatched for the slopes of partial tracks, and an
additional minus sign thrown in (this all due to different coordinate systems).  Still, a quick comparison allows
to conclude that the track reconstruction seems to mis-reconstructs the tangent of the scattering angle from
region 2 by -0.0032 +/- 0.0003.  That would be a systematic 0.2 degrees offset at 7 degrees (geometries are identical,
note on octants above), or 5.5% on Q2 (*).  I have a feeling we're about to find out some interesting stuff here...

(*) We never expected to get a valid Q2 from tracking data alone.  


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4801 b9700a98-9937-0410-b777-f318d6eb5bff