Just a simple perl script that takes in a detector name (from the md_data_view, so md's and bkgd detectors), slug range, and regression scheme, and outputs the IN, OUT, NULL, and PHYSICS asymmetries for that range. This functionality essentially already exists in the QwDBGUI, but this is a nice tool if you don't care about the plot but rather just in values and want them quickly.
You'll likely need to edit your connection information at the top of averages.pl in order for it to work, but I've tested it on the ifarm and it works fine if you put host=qweakdb.jlab.org and keep all of the other information the same.

Editting it to run on the lumi tables would be trivial if you'd like that as well; beam and slope tables would be a bit trickier.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4806 b9700a98-9937-0410-b777-f318d6eb5bff