This is an initial commit of the run1pass5 and run2pass5 config directories for the slugfile creation. A version of the run1pass5 directory has existed on the group disk but outside of version control for a while. I wanted to get a copy in svn so we can manage it a bit better. The run2pass5 directory is just a copy of the run1 one for the time being.
I have also inserted the regression class variables into the leaflists. These include variables like mdallpmtavg, dipoles, and the newly added mdNpmtavg variables.

The run_makeslug scripts will need to be editted to include the proper directories once we begin analysis, as well as directory references inside the MakeSlug* scripts.

The MakeSluglet script that we intend to use in the continuation of pass5 has not been propagated down into these directories, as we still have to test it with the batch submission on the farm, but once it has been cleared we can copy it down into them.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4830 b9700a98-9937-0410-b777-f318d6eb5bff