Commit of some local changes to the 3.04 version of the analyzer checked out onto the group disk.
run_makesluglet is a copy of the ppMakeSluglet script that was used for the Run2Pass1 analysis: renamed for consistency and changed to be more readable. The script also, for now, will execute the MakeSluglet_lrb.C script in addition to the standard MakeSluglet.C script, much in the same was as the MakeSlug_lrb.C script was used for the Run1Pass5 analsyis of the 25% dataset.

The run_makesluglet script will need to be edited as soon as we have the correct paths chosen.

I've done some early tests of this script and it was working fine for both unregressed and regressed ROOTfiles for one run.

The script is called the same way that the old ppMakeSluglet script was:
$ run_makesluglet RUN_NUM

Note to self or others: by pulling out the input file tree name into the required parameters one can eliminate the need for two separate scripts for regressed/unregressed Sluglet files. 


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@4867 b9700a98-9937-0410-b777-f318d6eb5bff