More local changes to the 3.04 version of the analyzer. This version now creates a Sluglet file for EACH RUN SEGMENT as opposed to one file for each run. We need to do it this way because the batch jobs sometimes split a single run into two jobs, so the Make_Sluglet.C script wouldn't have access to all of a run's segments all of the time. Additionally, making a file for each sluglet allows us to better control exactly which segments get put into the Slugfiles.
For instance, if a certain segment has "few" good events, it will still have a sluglet created for the unregressed results, but perhaps not the regressed ones. We would then lose a 1:1 correspondance of events between unregressed and regressed slugfiles.

For now the creation of regressed sluglets has been disabled by commenting out the appropriate lines in run_makesluglet. There's probably no reason to bother generating those files, because it only reduces the file size per segment from ~70Mb to ~55Mb. The unregressed files on the other hand, go from ~9Gb (when did they get that big???) down to ~210Mb.

the run_makesluglet script now takes in two arguments so that is can be called properly as a post-runlet job by qwbatchsub.pl. The arguments correspond to the run and segment numbers:

$ run_makesluglet RUN_NUM SEG_NUM

I've tested this script again on a test run using run1pass5 ROOTfiles for run 9720. The only issues it has is finding the mdNpmtavg variables because they didn't exist at that time.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@4868 b9700a98-9937-0410-b777-f318d6eb5bff