run_makesluglet just has a simple change of where it would look for lrb_rootfiles if we asked it to create regressed sluglet files. We do not do this now.
run_makeslug and run_makeslug_unreg were edited to use a leaflist appropriate for stitching together sluglet files instead of basic qwparity ROOTfiles. Our batch jobs do not currently call either of this scripts but rather are tools to be used in the future when creating slugfiles from the sluglets.

leaflist_run1pass5_sluglet.txt is the new leaflist appropriate for sluglet files as input to the MakeSlug.C script: we needed to change variables with '.' characters in them to be '_' characters.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@4947 b9700a98-9937-0410-b777-f318d6eb5bff