NOTE:  this is on branch release_3.04.
Merged the changes from trunk revision range 4906:4959 into branch
release_3.04.

This picks up one change in qwbatchsub.pl, some reordering of the
QwDBGUI lists, and some MD and injector mapfile changes.
The leaflist_run1pass5_unreg.txt was also changed to add some
injector variables to the slugfiles.

There are some other Extensions changes in the Compton and Transverse
subdirectories, but they are incidental to the main purpose of this
update.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@4960 b9700a98-9937-0410-b777-f318d6eb5bff