NOTE:  This is on branch release_3.04.
Modified the qwbatchsub to include a new option, "--end-script"
which will specify the name of a script to be run at the end of
the job.  This is intended to be used for rootfile staging.

Also added the capability to use a hidden file ".farm_os_name"
or an option "--farm-os" to specify the "OS name" variable to
be used in submitting the farm job.
The OS name defaults to "linux64" if neither of those are specified.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@4961 b9700a98-9937-0410-b777-f318d6eb5bff