NOTE:  This is on branch release_3.04.
Modified the official_run1pass5 and regression_run1pass5.sh to use
the qw_run1_pass5 database, and added the "end-script" call to
the outputstage_r1p5 script.

The outputstage_r1p5 will copy the histogram ROOT files for all runs
and the tree ROOT files for every fifth run to the silo.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@4962 b9700a98-9937-0410-b777-f318d6eb5bff