rootClass.h: added color printing ability infoDAQ.C: additional strips that can be masked (on top of the strips masked in DAQ) from the asymmetry fit;          also, the call of this function sets a Bool_t(maskSet) which is checked in the beginning of other functions that use mask values. comptonRunConstants: addition of variables/constants needed for above expAsym.C : now the infoDAQ() function is invoked conditionally based on whether the maskSet function is set; this makes the different files capable of running independently, and not one after another.
fileReadDraw.C : the addition of maskSet has happened; the laserCyc Accum and Scaler plots were not happening for multiple planes, hence for now, they  have been modified to do just the first plane. It will be debugged later.

asymFit.C : the functionality of ignoring a few strips, despite being in the to-be-fit graph has been implemented using TF1:RejectPoint(), 



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4963 b9700a98-9937-0410-b777-f318d6eb5bff