NOTE:  this is on branch release_3.04.
Modified the official_run1pass5 script to determine the QWANALYSIS
variable based on the path to the script (to make it easy to
support both CentOS6 and CentOS5 installations).

Modified the paths to linRegBlue to be derived from the QWANALYSIS
variable in the two regression control scripts; this should also
make sure that the proper linRegBlue is called from CentOS5 or CentOS6
machines.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@4999 b9700a98-9937-0410-b777-f318d6eb5bff