NOTE:  This is on branch linRegBlue_v7.
The LRB script which produces the files to be uploaded to
the DB needed some small changes to work with the root
installation on the CentOS6 machines.

It was verified to produce the same output file as the original
version on the CentOS5 machines using the 'std' output of 
runlet 10101.000.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/balewski.correlator@5002 b9700a98-9937-0410-b777-f318d6eb5bff