Imp: the asymFit.C uses the charge normalized accum counts to determine the compton edge. This is the only procedural change in this revision.
The purpose of the current update was to modularize the macros and improve the structure. This could eventually be used for event mode and scaler mode analysis with least modifications. Some variable names have been changed, while some have been defined now in the comptonRunConstants.h file

Due to the modularity now, the data corresponding to PWTL2 as well as scalers gives out the parellel corresponding information. Earlier this was not uniform, some information was not being generated for the PWTL2 and scaler data



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5026 b9700a98-9937-0410-b777-f318d6eb5bff