NOTE:  This is on branch release_3.04.
This corrects the mistake in filling the {beam,lumi,md}_errors
tables in the DB.  The errors tables are now filled with the
numbers of patterns which had that error condition for that
channel.  Note that error conditions are not exclusive; the
sum of error counts can be larger than the total number of
patterns.

Also, this adds some code to fill the general_errors table
with the blinder error counters.  Note that the blinder errors
are exclusive, and the sum of these should be exactly the number
of patterns.

This has bee tested against r5029 of release_3.04 for run 10101
and found to produce the same results.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@5032 b9700a98-9937-0410-b777-f318d6eb5bff