in the header file, the fStripsRaw for each data type was changed to double from int
in the script, the static_cast was made double in the scale function.
the indenting in this script was messed up due to opening in another editor, It is okay now but as a result a lot of lines appear to have been edited in this revision, this is not true. 



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5053 b9700a98-9937-0410-b777-f318d6eb5bff