Added a framework for dynamically loading "plugin" libraries/macros to be processed.
One specifies as a parameter the --macros blah blah2 options and they
will be processed in that order.

Libraries must already be compiled as a shared dynamic library (*.so)
and stored in the path specified by --macros-path.

Users are encouraged to see $QWANALYSYS/Extensions/Macros/Compton/beamline
as an example of how to do this (of course, that doesn't exist now, but
a commit is not too far away ;)

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5072 b9700a98-9937-0410-b777-f318d6eb5bff