NOTE:  this is on branch release_3.04.
The modulation analyzer had a double-application of the ramp
voltage conversion to angle, as it is done inside the engine,
and then was done again in the modulation analyzer.  This
results in a "ramp angle" which is scaled down by a factor
of 10.

The modulation analyzer now uses the ramp as reported directly
as in units of degrees, which is what is provided by the engine.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@5148 b9700a98-9937-0410-b777-f318d6eb5bff