NOTE:  this is on branch release_3.04.
Merged changes from revision range 4959:5075 on the trunk into
the branch.  Most of the changes are related to the Compton
analysis, but there are some DBTree changes and one change to
the Tracking engine.

This is an intermediate merge as I try to bring the branch and
trunk into consistency.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@5155 b9700a98-9937-0410-b777-f318d6eb5bff