NOTE:  this is on branch release_3.04.
This is a merge from trunk into the branch as of revision 5192.
The branch and the trunk are now completely consistent.

Here's what I did:
Starting with the head of the branch, merged the differences between the
trunk and branch as of revision 5169:
     svn merge https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@5169 https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5169 .
Then merged the remaining changes on the trunk from 5169 to 5192:
     svn merge -r 5169:5192 https://qweaksvn.jlab.org/repos/QwAnalysis/trunk .

No conflicts needed to be addressed.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_3.04@5193 b9700a98-9937-0410-b777-f318d6eb5bff