Newest release of QwDBTree.
MAJOR CHANGES: The run variable has now been renamed run_number to match the
syntax in the database. I'm no longer going to change any of the naming syntax.

MINOR CHANGES: I have added a run_number_decimal variable which is basically
run_number + segment_number/20. I did a quick check in the database and I don't
think there are any good runs with segment number greater than 20. I will look
more into this later.

I also added a sign_correction variable into the database. It uses the
prescription laid out here:
https://qweak.jlab.org/elog/Analysis+%26+Simulation/619. I checked the logic
for a few cases but I encourage you to check this variable.

I also added the wien_slug variable which is what is commonly referred to as
the 'wien number'. This variable is *NOT* correct in the database yet, so
please do not rely on it.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5409 b9700a98-9937-0410-b777-f318d6eb5bff