Edited this file so that it works with both data ROOT files and ROOT files for qwsimtracking.
Major changes include:
- adding in the cross section as a weighting for the histograms
- creating if statments that allow only the needed parts of the 
  code to be used for data files or files from qwsimtracking
  These come in because the reconstructed ROOT flies do not 
  have a main detector or trigger scintillator branch

On thing to note is that, because the simulation only have one 
octant with tracking chambers in them this means that all of 
the plots that are of package 2 are empty - which when running
this macro will give a small complant at the end due to a log
scaling on one of the plots. This however does not effect the 
results - so don't worry about it.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5518 b9700a98-9937-0410-b777-f318d6eb5bff