This is a program I wrote a year ago and just updated today. Eventually it will read in a table of slug data and compute the error-weighted average and uncertainty.
I know there are other tools out there to do this exact thing,
but I wrote this a year ago and it compiles and works, so I
will upload it here and in case anyone else wants to use it.

Currently this program just takes dummy data, but the structures
are there to do more. I will continue to improve this over the next
few days as we wait for the official plots to be produced.

To compile:
g++ -Wall slugavg.cc -Iinclude `root-config --ldflags --cflags --libs` \
-o testavg src/slugcalc.cpp src/slugveccalc.cpp

Class structure and definitions:
SlugCalc is a class that takes slug information and
correctly calculates the average data information.

SlugVecCalc it is a vector of SlugCalcs.
Therefore, one can have a vector where each element is a SlugCalc
object that can computer weighted averages and errors independently.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5532 b9700a98-9937-0410-b777-f318d6eb5bff