This changes the search order of directories during creation of new SET_ME_UP files, in order to fill the environment variables:     MYSQL_INC_DIR     MYSQL_LIB_DIR     MYSQLPP_INC_DIR     MYSQLPP_LIB_DIR     BOOST_INC_DIR     BOOST_LIB_DIR
In the old search order, the directory referenced by
the currently set variable was searched first; the
new order will search that directory last.

This should make setting up after a major update
easier, as if the best matching directory has changed,
the script will not default to the old value.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5581 b9700a98-9937-0410-b777-f318d6eb5bff