These are the long-anticipated fixes for the BMOD ramp decoding.
The channel map "qweak_modulation.18894.map" was renamed
"qweak_modulation.18894-.map", to be used for all later runs
also.

Several pedestal ranges were added based on the Elog entry
Analysis&Simulation/1027.
The run ranges could probably (should probably) be opened up
so there are not gaps, but for now that only run range that
we care about is the one which begins with run 18894 (the last
two slugs or so of run2).

The run ranges have not been opened up, because I am not sure 
which ranges should be extended in which direction.  If we decide
to do a full pass6, this may be something we want to look at
before or during the analysis pass.
Also note that some of the speacial purpose runs may fall in
the gaps; if we need to rerun one of those, we probably want to
check the ramp response and fix the pedestal before a large
replay.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5606 b9700a98-9937-0410-b777-f318d6eb5bff