New command line flag --QwRayTracer.initial_momentum <float> (defaults to 1.25 GeV).
This is the value (in GeV) of the initial momentum guess for the Newton's method that
determines the scattered momentum.  We try to keep this higher than the elastically
scattered momentum (after finds by Jie).  For Moller scattering this may need to be
radically lower.

One remaining idea is to build in the logic such that a negative value would be
interpreted as turning on the estimator based on the HDC angle and the specified
rest mass of the target.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5689 b9700a98-9937-0410-b777-f318d6eb5bff