Once upon a time there was a group of physicist who were analyzing tracking data.  These people were fed up with having to search for data so one day a graduate students stuck in a cubical decided to take it upon them-self to create away for this to happen.
From the depths of disrepair of this daunting task came new hope as and idea 
came to mind, ROOT, while it would be a big task to take on, it was the way to go.  
From the forest with the many ROOT that kept tripping up this student, came 
another with a helping hand to navigate through.  After a long trek and many hours, 
sweat, and headaches, came ValerianROOT away to relax and lose the tension 
that had taken refuge in shoulder hunched over a computer. 

ValerianROOT was (will be if all goes well) greeted with great cheers as now 
tracking down of the monster that seamed to be allusive was no longer, While 
a ways a way the path was clearer.  ValerianROOT had come to the rescue!!!

This addition to QwAnanlysis is responsible for getting a ROOT file out for 
all the information gather in text files from the tracking website.

**Public service anncment: this is not 100% functional or optimal.
It is getting uploaded since a certian wildcards tend to show up 
where they shouldn't be for a certian person.  There is to much work
done for it to get lost.**

Please wait for a green light commit to insure safe passage.  This is
not safe yet. :( but will be soon


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5749 b9700a98-9937-0410-b777-f318d6eb5bff