Okay now where were we in the tail of ValerianROOT...
Oh yes, graduate student long hours of fighting with a computer...
common story.  But the dragon has been slain and the prince rescued
from the tower.  We have, (drum roll) ValerianROOT!!

This is the "version" of ROOT (since qweak didn't have enough of them
already ;P) to look at and manage all the tracking data at once.  Like
the title implies is calm and relaxing (okay you are right not always,
we all know how ROOT is).

The README file is the indepth place to look at how this works, outside
of unicorns and magical goblins and the gremlins feed after midnight
wreaking havoc.  But in a nut shell it:
  - Takes all the text file output from the tracking website and makes it
    into ideal text files
  - Reads those text files into definded structures
  - Takes those strutures and makes a wonderful ROOT tree out of it
  - Debugging in the process if need (for when one adds to ValerianROOT,
    cause they might be dragons...)

There is a macro that plots Q2 vs Run number for a list of runs, (run list
can be found and stored in runlist).

To just use ValerianROOT, run the Compile.sh script.
To run and MAKE a ValerianROOT tree, use the Run_it_al.sh script - this
and what it is connected to etc will be useful when adding on to ValerianROOT.

Consult the README for more info.

And of course enjoy!



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5780 b9700a98-9937-0410-b777-f318d6eb5bff