These mapfile changes were performed in order to prepare for the 3-pass replay. Updates include:
*removing md3 and md7 from the qwk_mdallbars, mdallbarsavg, and mdallpmtavg variables. Weights in the regression.map files were appropriately changed to reflect the reduced channel count.

*new beamline_eventcuts files for 3-pass aluminum running. A reduction in the current cut from 100 uA to 20 uA was done since we ran aluminum  at 60 uA.

We (Jim and Scott)  will need for these changes to be merged into the analyzer that is checked out on the ifarm before beginning the replay. We have not made any changes to the linregblue config files, because we don't think we will need to...seeing as though the md7 removal was done here instead, and is still a channel that we want to go through the usual analysis procedures.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5788 b9700a98-9937-0410-b777-f318d6eb5bff