Its this version, that was used while making EDET ELOG 607.
evaluateAsym.C:
deleted commented out lines from past which are not needed anymore
Noise subtraction needs to be done in expAsym.C, so I need the helicity disentangled counts outside the evaluateAsym.C, hence taking these operations outside of this function. Now I'll pass on the counts itself, which can then be further processed in this function

expAsym.C:
deleted commented out section of the method where I was throwing out those laser cycles which had a beam trip.
added explicit noise subtraction.
Removing the 'if(!mask[p][s])' method of skipping the masked strips, because few runs intermittently have wrong entries in the configuration tree (reason unknown yet). But because we already know that strips 2,6 and 20 were masked, we will not use them while finding pol% from asymmetry, but otherwise, there's no harm in reading and writing for all the strips including the masked ones. So in this new approach, the 'expAsym.C' will process all strips and write the corresponding numbers. I anyways have a check on the denominator being zero, so the masked strips can't cause a crash. 
Removed the 'activeStrips' implementation, now all strips, if read in from somewhere will be from a file, so this will never be needed

comptonRunConstants.h:
added final beam-off variables that may need to be used across functions

edetLasCyc.C:
changed the void into an integer function returning 2, This can be used to check a graceful exit of the execution.

writeToFile.C:
commented out the cout statements of the creation and filling of files. I don't find them relevant anymore

infoDAQ.C:
now the list of 'additionally masked strips' is the basic list of masked strips which will be used to skip the strips from the asymmetry determination. This list now has 2,6,10 and 20 (keeping only plane-1 in mind)

rhoToX.C:
changed the pol3 fit into a pol5 fit to make it compatible with runlet based analysis

asymFit.C:
it has a 6 parameter fit (pol5) to the QED asymmetry w.r.t distance, instead of the earlier 4 parameters fit. This is done just for the sake of matching with the runlet analysis.

polCERunRange.C:
changed some variable names, adding the capability to plot from a file if the runlet pol% comes in the format similar to lasCyc

fileReadDraw.C, qNormVariables.C:
commented out the if(!mask[p][s]) statements, because the mask[p][s] variable is not being populated anymore from this version onwards. I will, if needed have to use a different mechanism to ignore strips

batchsub.sh, rootlogon.C, batchsub_header1.xml : trivial changes like changed options and comments






git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5827 b9700a98-9937-0410-b777-f318d6eb5bff