Ability to turn on/off the dropping of the hit with worst residual in track fit with the argument --QwTracking.R2.DropWorstHit <bool>.  Default is off.
The way this had been implemented was not optimal.  Imagine the following possible
partial tracks in region 2:
(1) pt1 with chi^2/ndf = 1.5
(2) pt1 after dropping worst hit with chi^2/ndf = 1.2
(3) pt2 with chi^2/ndf = 1.6
(4) pt2 after dropping worst hit with chi^2/ndf = 1.1

Before this revision, the code would evaluate option (1), then evaluate option (2).
Because (2) has a chi^2/ndf that is 80% or smaller than the chi^2/ndf of (1) it 
would keep track of option (2) as the best option.  Then it it reaches option (3),
which is discarded because of the higher chi^2/ndf, and this automatically also
discards option (4) (which is the one we would have expected it to pick).

After this revision, the code will pick option (1) by default (i.e. with the argument
--QwTracking.R2.DropWorstHit false).  It will pick option (4) when explicitly the argument
--QwTracking.R2.DropWorstHit true.

Someone will have to check the impact of dropping these worst hits, but at least for now
it is switched off by default.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5836 b9700a98-9937-0410-b777-f318d6eb5bff