In the dark ages we evaluated some different structures for how the QwEvent class stores the tracking information.  We evaluated a regular TClonesArray, a static TClonesArray (improved memory footprint and better cache hits), and inheritance of STL containers through VQwTrackingElement.  None of these worked as well as just having STL containers as class members.  ROOT now supports those well enough.
In any case, we still had a lot of legacy code that used preprocessor macros
to pick the correct code.  That's been weighing down some things we want the
code to do.  So, I finally stripped out the unused bits (and bytes).

(Verified on run 15121 that nothing changes if you remove code that doesn't
get compiled anyway.)


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5838 b9700a98-9937-0410-b777-f318d6eb5bff