weightedMean.C: if the weightedMean of the denominator of the asymmetry comes negative, the macro will now stop execution returning -1.0 with a red warning message.
evaluateAsym.C : It now implements deadtime calculation also. It also has some modified variables to make them brief and more intuitive

expAsym.C: major addition; Implements deadtime correction. Currently, it reads in two files that have the p0 and p1 order corrections for deadtime correction to be applied for a given run; this method of reading in the deadtime correction factors may change, but the the final correction factor is passed on to evaluateAsym.C; that can be left unchanged.

weightedMean.C: it was allowing for special cases of the weighted mean being negative. This was a wrong provision. Now it exits the execution when found with a negative weighted mean of yield

comptonRunConstants.h: it has additional variables for deadtime correction. Some variables have been moved from fileReadDraw.C to here because of their repeat usage in other macros. Also added skipCyc variable. 

asymFit.C: Having shown in elog 611 and 612 that the runlet and laserCycle methods are essentially the same, I've removed references of runlet data or runlet method that was left here with the possibility of comparing the results in two cases. The maximum number of iterations have also been increased to 10000, this helps the case when the fit routine abruptly terminates because of not finding the fit result in the default no.of iterations

rootClass.h: added TVirtualFitter.h to the list of headers added

edetLasCyc.C: now does not include the fileReadDraw.C macro during regular execution of polarization extraction package.

infoDAQ.C: major addition, it now reads in the slow tree and writes the IHWP and similar variables into a file. 

batchsub.sh and batchsub_header1.xml have trivial changes to submit new set. 

fileReadDraw.C: variables NDF and chiSq have now been moved to comptonRunConstants.h



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5852 b9700a98-9937-0410-b777-f318d6eb5bff