evaluateAsym.C: removed the reference of skipCyc, it was a left over from a previous attempt at implementing something
weightedMean.C:
removed the allowance of getting a negative value for the weighted mean of any quantity. it was allowed in the code with a warning message.

asymFit.C: major
the file read in has been made more robust and correct. The residulas were getting calculated wrong. That is fixed in this revision. This version also plots the background asymmetry and finds a linear fit to it. The linear fit values of residuals and background asymmetry are now written into a file to easy access across runs

comptonRunConstants.h:
added variables for plotting background asymmetry, in the asymFit.C

The current directory also has a modified version of expAsym.C , which is modified in a way to accept "only clean" laser cycles, and throw a laser cycle if it has a beam trip. We are currently evaluating which mechanism is best and hence not putting this as ansvn revision, because we may have to undo this.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5858 b9700a98-9937-0410-b777-f318d6eb5bff