for all macros, the reference to rootClass.h was changed to "rootClass.h" instead of <rootClass.h>, since it is a local file
asymFit.C: 
made the 'theoreticalAsym' a 2 parameter function. Till now it was a 3 parameter function with the first one (effective strip width) as a fixed parameter. Doing this did help for few runs eg. 23355; but still not for several other runs eg. 23415

comptonRunConstants.h: 
now the effective strip width variable is a constant defined in this header instead of the asymFit.C macro

batchsub.sh: 
added a check on whether the rootfile even exists or not, before submitting the xml file for that run. 



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5872 b9700a98-9937-0410-b777-f318d6eb5bff