infoDAQ.C Turned off debug. Added a red print if the set and readback value of the IHWP1 is different. The macro now exits if the IHWP or the edet position was changed during the run by looking at the ratio of rms_<readback>/<readback> of ihwp1, ihwp2 and edet-pos variables. Reading in beam energy and writing that into the beamProperties file output. This read in number is not yet being passed on to rhoToX.C macro. Changed the name of output file to *_beamProp.txt instead of *_beamProperties.txt.
asymFit.C:
exits if the infoDAQ exits with an error. 
The polarization fit is now in a loop which iteratively checks for the status to return 'SUCCESSFUL'. 
The numeric fitStatus is also added to the fitInfo file output.
This still is unable to avoid the 0.00 error of Compton edge (eg. run 23610, 23630) 

expAsym.C:
exits if the infoDAQ exits with an error

rhoToX.C:
changed the 'E' variable to eEnergy (electron beam energy). Now its not a constant and gets its value from infoDAQ.C macro as read in from the slow tree; though the value is still being set by hand to the old value




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@5873 b9700a98-9937-0410-b777-f318d6eb5bff